[DEV-4921] Move tests about fetch_jira_issues into right package
